tramp support: explicitly prohibit use of pager
Git is smart enough to detect that it runs inside emacs and it don't
need to run pager. But when git is started by tramp via ssh connection
it still runs it and waits at "terminal is not fully functional"
prompt with no way to press return key. It's possible to achieve the
same result with setting PAGER environment variable to "cat", but that
would require converting everything to use magit-shell instead of
magit-run, because tramp doesn't transfer environment to remote host.

NOTE: this change requires git at least 1.5.3rc8, where --no-pager
option was introduced.