Use pipe rather than tty for connection with git.
When remote updating, git may use ssh that may ask for a password or
passphrase. ssh used to be on a tty, and so believe that it could ask
the passphrase directly to the user. It now know that it can't, and
will ssh-askpass (if available) to do the asking.

It would be probably better to detect the passphrase asking, and to
use Emacs to ask for it.